[Android] MRP-556: Refactoring & code clean up.
