+++ plugins/quick-brick-adobe-primetime-auth/android/src/main/java/com/applicaster/adobe/login/pluginconfig/PluginDataRepository.java
@@ -1,6 +1,7 @@
-package com.applicaster.adobe.login;
+package com.applicaster.adobe.login.pluginconfig;
 
-import com.applicaster.adobe.login.model.PluginConfig;
+
+import com.applicaster.adobe.login.pluginconfig.model.PluginConfig;
 
 public enum PluginDataRepository implements PluginRepository {
     INSTANCE;
